requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==24.1.0
starlette==0.38.6

# Scraping
google-search-results==2.4.2
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.27.2

# Data Processing & Export
pandas==2.2.3
openpyxl==3.1.5

# Security & Rate Limiting
slowapi==0.1.9
python-dotenv==1.0.1
itsdangerous==2.2.0

# Async utilities
asyncio-throttle==1.0.2
